球场状态的数据页优化
球场状态的数据页的cell自定义方式显示，以及地图中的添加模拟的组GPS显示点160426

+++ xygolf/customView/EachHoleMoreInfo.m
@@ -205,7 +205,7 @@ typedef void (^CustomCompletionAnimationBlock)(BOOL finished);
         [_moreHanleView addSubview:eachView];
     }
     
-    NSLog(@"subviewCount:%ld",_moreHanleView.subviews.count);
+    NSLog(@"subviewCount:%ld",(unsigned long)_moreHanleView.subviews.count);
     //
     [contentView addSubview:_moreHanleView];
 }
